requirements.txt
slack_sdk>=3.0
slack-bolt
Flask
python-dotenv
langchain